Uniform formatting
I began to organize the image folders, added a few images, and began working on some actual content for the site. I also converted all spaces to tabs, which is why there are so many changed files with this one.

On top of that, I made it so longer descriptions are cut short for projects, in case either me or Kara gets overzealous and writes too much content.