Populated with content
It's far from perfect, but I now have the main body of content filled out to a satisfactory level. I plan on getting feedback from several people soon to improve it, but the first priority is watermarking images and limiting the display of artwork to six with the option to show more with pagination.