"Load More" projects button
Found a guide that did a decent job explaining how to lazy load in more projects with React, so I implemented and improved the offered methods. Also went back and added semicolons to all the files, because not having them scares the shit out of me.

The skills section also fades in now, to fit in with the flow of the rest of the site. Next step is to knock off the last few remaining TODOs and improve SEO for the search console, if Google would hurry up and index the site.