Watermark images before runtime
Now the user's device will never need to bear the burden of watermarking images, because the server will do that on every code deploy automatically. I'm ignoring the watermarked images in the static folder because I don't think the server will need them, but I can come back and change that if necessary.

It would be more effective to only watermark images locally before committing to the server, but making it automatic makes it less prone to user error, so I'm willing to take the hit to build times/cost.