Pushing all local changes
I haven't worked on this in a while so I'm going to push all the changes I have before we attempt to take the site in an entirely new direction. I will be creating a branch with this code once pushed to be able to easily revert/retrieve the code if I need to do so in the future.