Removed "undeclared vars" rule
Every single occurrence of a violation of the rule was because the variables were declared in the global scope, and the codebase is still very easy to navigate and keep track of. So I turned off the rule, and the global comment at the top of several files is no longer necessary. This will allow me to focus more on real issues instead of imagined ones.