CORS compliant
The JS Paint app now cooperates with my Nginx configuration for the server, and the application is opened directly from the file manager. The images are now also faded around the edges and there are plans in place to generalize the code for potential future use cases. The top bar has not been started yet, that is the next step.